разграничение пользователей

+++ Pages/Blog/Delete.cshtml.cs
@@ -1,9 +1,12 @@
-﻿using Microsoft.AspNetCore.Mvc.RazorPages;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.AspNetCore.Mvc;
 using RadioAmateurHelper.Data;
+using System.Security.Claims;
 
 namespace RadioAmateurHelper.Pages.Blog
 {
+    [Authorize]
     public class DeleteModel : PageModel
     {
         private readonly ApplicationDbContext _context;
@@ -18,7 +21,7 @@ namespace RadioAmateurHelper.Pages.Blog
         public IActionResult OnGet(int id)
         {
             var post = _context.BlogPosts.FirstOrDefault(p => p.Id == id);
-            if (post != null)
+            if (post != null && CanDeletePost(post.AuthorUserId))
             {
                 _context.BlogPosts.Remove(post);
                 _context.SaveChanges();
@@ -29,6 +32,22 @@ namespace RadioAmateurHelper.Pages.Blog
             }
 
             return RedirectToPage("/Blog/Index");
+        }
+
+        private bool CanDeletePost(string? authorUserId)
+        {
+            if (User.Identity?.IsAuthenticated != true)
+                return false;
+
+            if (string.Equals(User.Identity.Name, "Leka-07@bk.ru", StringComparison.OrdinalIgnoreCase))
+                return true;
+
+            var settings = _context.SiteSettings.FirstOrDefault();
+            if (settings?.AllowUsersToDeleteOwnPosts != true)
+                return false;
+
+            var currentUserId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            return !string.IsNullOrEmpty(authorUserId) && authorUserId == currentUserId;
         }
     }
 }
